+++ packages/convostack-backend-express/src/index.ts
@@ -1,4 +1,1 @@
 export * from "./backend";
-export type {
-    IConversationEventServiceOptions, IConversationEventServicePubSubEngine, IConversationEventServiceCache
-} from "./services/conversation-event.service";