+++ packages/convostack-models/src/graphql-context.ts
@@ -1,0 +1,5 @@
+import {IUser} from "./storage-engine";
+
+export interface IGQLAuthContext {
+    user: IUser | null;
+}