+++ examples/be-example-express-sqlite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # be-example-express-sqlite
 
+## 0.0.26
+
+### Patch Changes
+
+- Move caches to models
+- Updated dependencies
+  - convostack@0.0.29
+
 ## 0.0.25
 
 ### Patch Changes