+++ packages/convostack-agent-sse/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @convostack/agent-sse
 
+## 0.0.26
+
+### Patch Changes
+
+- Move caches to models
+- Updated dependencies
+  - @convostack/agent@0.0.26
+  - @convostack/shared@0.0.26
+
 ## 0.0.25
 
 ### Patch Changes