+++ packages/convostack-frontend-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @convostack/frontend-react
 
+## 0.0.27
+
+### Patch Changes
+
+- Move caches to models
+
 ## 0.0.26
 
 ### Patch Changes