+++ packages/convostack-langchain-memory/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convostack/langchain-memory
 
+## 0.0.27
+
+### Patch Changes
+
+- Move caches to models
+- Updated dependencies
+  - @convostack/shared@0.0.26
+
 ## 0.0.26
 
 ### Patch Changes