+++ packages/convostack-storage-engine-prisma-postgres/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @convostack/storage-engine-prisma-postgres
 
+## 0.0.27
+
+### Patch Changes
+
+- Move caches to models
+- Updated dependencies
+  - @convostack/models@0.0.26
+
 ## 0.0.26
 
 ### Patch Changes